# GS: Enable small framebuffers at 1x