### Description of Changes
- Stops the texture cache from downscaling targets when used as a source
- Enables smaller framebuffers even when rendering at 1x (previously only enabled at 2x+)
- Adds some optimizations to move sampling to hardware when possible (minor perf improvement on GT4)

Note: I did this by editing the vertex and fragment shaders to support rescaling normalized coordinates.  An alternative way to do it would be to just allocate full scale textures in this case.  Preferences?  (sampling the target directly mentioned below would need this approach)
Note2: This changes the renderer's interpretation of a texture whose `size / scale` does not match `tw` and `th` from "stretch to fit" to "use a 1-1 pixel scale".  Not sure if this breaks anything?  If it does, maybe those things should have set their scale properly...

### Rationale behind Changes
- Improves speed of Burnout at 1x on integrated GPUs with low memory bandwidth due to its access patterns making TC convert framebuffers to and from depth formats constantly (by reducing the size of those framebuffers).
- Improves the image quality of Burnout when upscaled to be the multiplier you selected and not that times 80% or so then nearest-neighbor upscaled back to 100%
- Should make removing the extra copies we do when using a target as a source a bit easier (future improvements, not included in this PR)

Example at 2x upscale:
<img width="640" alt="ExampleA" src="https://user-images.githubusercontent.com/3315070/149651911-0406adae-97f1-412f-858f-14c5726ac487.png">
<img width="640" alt="ExampleB" src="https://user-images.githubusercontent.com/3315070/149651913-6539a19d-3571-4745-a346-782b7c44c537.png">

### Suggested Testing Steps
Test 1x and 2x upscale for lots of games