@refractionpcsx2 2022-01-23
The second picture definitely looks much better, the first picture is very blurry