@tellowkrinkle 2022-01-23
Not sure blurry is how I'd describe it, nearest neighbor is really garbage for non-integer upscales
Either way it looks terrible