@refractionpcsx2 2022-01-23
I'm afraid this doesn't work so well with mipmapping, as seen here in ratchet

![image](https://user-images.githubusercontent.com/6278726/150703493-55ebcc86-6d75-4042-9aa4-e546eabba41b.png)

should look like

![image](https://user-images.githubusercontent.com/6278726/150703515-dba6123f-fa94-4c17-84b4-25c96cce33be.png)

GSDump:
[ratchetmips.zip](https://github.com/PCSX2/pcsx2/files/7921747/ratchetmips.zip)