@tellowkrinkle 2022-01-30
Mipmap issues should hopefully be fixed