@refractionpcsx2 2022-01-30
That's fixed, but there is another issue, it looks like the "half_right" code no longer works for Baldurs Gate 2 or Justice League. Seems OK on master.

![image](https://user-images.githubusercontent.com/6278726/151700016-71e9e8c6-933a-431b-a451-a14f3880d401.png)


Here's some dumps for you
[baldurgate2.zip](https://github.com/PCSX2/pcsx2/files/7965892/baldurgate2.zip)
[justiceslow.zip](https://github.com/PCSX2/pcsx2/files/7965893/justiceslow.zip)