@Dreadmoth 2022-01-31
Ratchet & Clank looks much nicer at any upscaled resolution 👍

-----
Videos seem to be broken in Baldur's Gate: Dark Alliance - this should be one of the intro logos:

![BGDA-brokenvideo](https://user-images.githubusercontent.com/8492684/151734002-ce4f6499-82c2-41bf-81ad-6e8ad6b727af.png)

Also affects the background in the main menu.
Occurs with all hardware renderers, native or upscaled.

[GS dump](https://github.com/PCSX2/pcsx2/files/7967868/BGDA_broken_menu_video_pr5387.zip)


-----
Champions of Norrath performance is significantly reduced (particularly on the main menu), without any noticeable change to image quality.


[Master menu](https://user-images.githubusercontent.com/8492684/151735540-4f0e4170-c3c1-4448-ab98-c11938e1c89b.png) (64 FPS)


[PR menu](https://user-images.githubusercontent.com/8492684/151735560-8b2a07bc-b35d-421d-b789-0b0eac60b5a4.png) (17 FPS)


[Master in-game](https://user-images.githubusercontent.com/8492684/151735581-a99bd8c3-1a67-455b-b440-85a1c2e38c89.png) (59 FPS)


[PR in-game](https://user-images.githubusercontent.com/8492684/151735585-7832302c-063f-478b-a757-33d0e42057f8.png) (27 FPS)

-----
Transformers has some kind of half-opacity fullscreen overlay which is upscaled on master, but appears to render at native resolution with nearest-neighbour scaling on this PR, resulting in the game looking progressively more pixelated as resolution is increased beyond 2x Native.

[Master 4x Native](https://user-images.githubusercontent.com/8492684/151733449-9b70b6b6-bdb5-49f7-ae5e-85fbd01ed322.png)

[PR 4x Native](https://user-images.githubusercontent.com/8492684/151733463-579bd1c2-43b6-4f13-bc48-6756d9446df6.png)

[GS dump](https://github.com/PCSX2/pcsx2/files/7967681/transformers_pixelation_pr5387.zip)