@tellowkrinkle 2022-01-31
Oh this PR is from when we had an issue with upscaling
Let me rebase and see if it's fixed for you