@tellowkrinkle 2022-01-31
Transformers and Baldur's Gate should be fixed
Champions of Norrath perf probably as well but not sure so please retest