@refractionpcsx2 2022-01-31
Perf has returned, no visible bugs