@DonelBueno 2022-02-18
This PR introduces a glitch in Gran Turismo 4, a darkened area at the bottom of the screen, as can be seen in the following screenshot:

![PCSX2_2022_02_18_18_33_50_735](https://user-images.githubusercontent.com/4239694/154765340-83f646d6-3f9c-48c8-8dff-2908f21a9efb.png)

It happens in any resolution above native and no change in settings fixes it.