@refractionpcsx2 2022-02-18
Can you provide a GS Dump?