@DonelBueno 2022-02-19
There you go:
[gs_20220219065418.gs.zip](https://github.com/PCSX2/pcsx2/files/8101885/gs_20220219065418.gs.zip)