@refractionpcsx2 2022-02-19
It works fine, you need to set Half Pixel Offset to "normal".