@DonelBueno 2022-02-19
Feeling kinda dumb right now x_x 
I thought it was a glitch because it didn't need that before this PR.

Thanks for the help.